Move normalization function to getEntityRecord selector
