+++ packages/core-data/src/test/resolvers.js
@@ -122,30 +122,6 @@ describe( 'getEntityRecord', () => {
 			1
 		);
 	} );
-
-	describe( 'normalizing Post ID passed as recordKey', () => {
-		it( 'normalizes any Post ID recordKey argument to a Number via `normalizeArgs` method', async () => {
-			const normalized = getEntityRecord.normalizeArgs( [
-				'postType',
-				'some_post',
-				'123',
-			] );
-			expect( normalized ).toEqual( [ 'postType', 'some_post', 123 ] );
-		} );
-
-		it( 'does not normalize recordKey argument unless it is a Post ID', async () => {
-			const normalized = getEntityRecord.normalizeArgs( [
-				'postType',
-				'some_post',
-				'i-am-a-slug-with-a-number-123',
-			] );
-			expect( normalized ).toEqual( [
-				'postType',
-				'some_post',
-				'i-am-a-slug-with-a-number-123',
-			] );
-		} );
-	} );
 } );
 
 describe( 'getEntityRecords', () => {
